rm index2
