+++ src/index.tsx
@@ -1,6 +1,6 @@
 /* @refresh reload */
 import { render } from 'solid-js/web';
-import { App } from './App/index2.tsx';
+import { App } from './App/index.tsx';
 import './index.css';
 
 const root = document.getElementById('root');
